requirements.txt
gradio
groq
python-dotenv
gtts
pydub
SpeechRecognition
ffmpeg-python
